obs-studio: Temporary fix due to cmake v4 change
The latest OBS Studio release doesn't currently build in the github
action due to a change in cmake version 4 behaviour. Cmake 4 doesn't set
the CMAKE_OSX_SYSROOT variable in the same way, which breaks SDK version
detection. Upstream has a fix, but until that fix is included in a
tagged commit, switch to a forked that contains the fix so that the
plugin can continue to build.